Split iDefinitions file.
