+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/ICastDefinitionBuilder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a12ddb87fd71c4629b1950456a5f8c94
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
