+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/IContextDescriptor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 206af15284fd34042a7b0a2c4342dfb7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
