+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/INodeDefinitionBuilder.cs
@@ -1,0 +1,20 @@
+using UnityEditor.ShaderFoundry;
+
+namespace UnityEditor.ShaderGraph.GraphDelta
+{
+
+    internal interface INodeDefinitionBuilder : IRegistryEntry
+    {
+        void BuildNode(
+            INodeReader userData,
+            INodeWriter generatedData,
+            Registry registry
+        );
+
+        ShaderFunction GetShaderFunction(
+            INodeReader data,
+            ShaderContainer container,
+            Registry registry
+        );
+    }
+}
