+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/INodeDefinitionBuilder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4efc727a2c954897bfde728c573d89d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
