+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/IRegistryEntry.cs
@@ -1,0 +1,11 @@
+namespace UnityEditor.ShaderGraph.GraphDelta
+{
+    public interface IRegistryEntry
+    {
+        RegistryKey GetRegistryKey();
+
+        // Some flags should be automatically populated based on the definition interface used.
+        // This whole concept may prove unnecessary.
+        RegistryFlags GetRegistryFlags();
+    }
+}
