+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/IRegistryEntry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 370ab2762e9414f2b8cf3ef0011c2cb0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
