+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/IShaderContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a70698ee028ae4b50b49077e2fcf19e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
