+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/ITypeDefinitionBuilder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a290d0a4ac394ac1871759491c0443a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
