+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/ReferenceNodeBuilder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74f7d1beaf3b04e1db9db02fe3d03655
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
