--- com.unity.shadergraph/Editor/GraphDeltaRegistry/StandardDefinitions.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5e80e323c66744d32a8a44e81eb4361e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
