+++ com.unity.shadergraph/Editor/ShaderFoundry/AssemblyInfo.cs
@@ -5,4 +5,3 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.ShaderFoundry.BlockSS")]
 [assembly: InternalsVisibleTo("Unity.ShaderFoundry.Editor.GraphicsTests")]
 [assembly: InternalsVisibleTo("Unity.ShaderGraph.GraphDeltaRegistry")]
-[assembly: InternalsVisibleTo("Unity.ShaderGraph.Test")]
